add namespace to listoptions
