+++ pkg/argoutil/resource.go
@@ -15,14 +15,7 @@
 package argoutil
 
 import (
-	"context"
 	"fmt"
-
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/types"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
 
 	argoprojv1alpha1 "github.com/argoproj-labs/argocd-operator/api/v1alpha1"
 )
@@ -48,12 +41,6 @@ func GenerateUniqueResourceName(instanceName, instanceNamespace, component strin
 	return fmt.Sprintf("%s-%s-%s", instanceName, instanceNamespace, component)
 }
 
-// FetchObject will retrieve the object with the given namespace and name using the Kubernetes API.
-// The result will be stored in the given object.
-func FetchObject(client client.Client, namespace string, name string, obj client.Object) error {
-	return client.Get(context.TODO(), types.NamespacedName{Namespace: namespace, Name: name}, obj)
-}
-
 // FetchStorageSecretName will return the name of the Secret to use for the export process.
 func FetchStorageSecretName(export *argoprojv1alpha1.ArgoCDExport) string {
 	name := NameWithSuffix(export.ObjectMeta.Name, "export")
@@ -61,14 +48,4 @@ func FetchStorageSecretName(export *argoprojv1alpha1.ArgoCDExport) string {
 		name = export.Spec.Storage.SecretName
 	}
 	return name
-}
-
-// IsObjectFound will perform a basic check that the given object exists via the Kubernetes API.
-// If an error occurs as part of the check, the function will return false.
-func IsObjectFound(client client.Client, namespace string, name string, obj client.Object) bool {
-	return !apierrors.IsNotFound(FetchObject(client, namespace, name, obj))
-}
-
-func FilterObjectsBySelector(c client.Client, objectList client.ObjectList, selector labels.Selector) error {
-	return c.List(context.TODO(), objectList, client.MatchingLabelsSelector{Selector: selector})
 }
